requirements.txt
Django>=5.2.3
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
Pillow>=10.0.0 